+++ fabric/src/main/java/de/maxhenkel/voicechat/intercompatibility/FabricClientCompatibilityManager.java
@@ -11,6 +11,7 @@ import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.HudLayerRegistrationCallback;
 import net.fabricmc.fabric.api.client.rendering.v1.IdentifiedLayer;
+import net.fabricmc.fabric.api.event.Event;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
@@ -25,6 +26,7 @@ import java.util.function.Consumer;
 public class FabricClientCompatibilityManager extends ClientCompatibilityManager {
 
     private static final ResourceLocation VOICE_CHAT_ICON_LAYER = ResourceLocation.fromNamespaceAndPath(Voicechat.MODID, "hud");
+    private static final ResourceLocation EARLY_JOIN = ResourceLocation.fromNamespaceAndPath(Voicechat.MODID, "early_join");
 
     private static final Minecraft mc = Minecraft.getInstance();
 
@@ -32,6 +34,7 @@ public class FabricClientCompatibilityManager extends ClientCompatibilityManager
         HudLayerRegistrationCallback.EVENT.register(layeredDrawer -> {
             layeredDrawer.attachLayerBefore(IdentifiedLayer.STATUS_EFFECTS, VOICE_CHAT_ICON_LAYER, this::onRenderVoiceChatLayer);
         });
+        ClientPlayConnectionEvents.JOIN.addPhaseOrdering(EARLY_JOIN, Event.DEFAULT_PHASE);
     }
 
     private void onRenderVoiceChatLayer(GuiGraphics guiGraphics, DeltaTracker deltaTracker) {
@@ -110,7 +113,8 @@ public class FabricClientCompatibilityManager extends ClientCompatibilityManager
 
     @Override
     public void onJoinWorld(Runnable onJoinWorld) {
-        ClientPlayConnectionEvents.JOIN.register((handler, sender, client) -> onJoinWorld.run());
+        // Higher priority to prevent this from not firing in case another mod throws an exception in this event
+        ClientPlayConnectionEvents.JOIN.register(EARLY_JOIN, (handler, sender, client) -> onJoinWorld.run());
     }
 
     @Override