+++ fabric/changelog.md
@@ -1,2 +1,3 @@
 - Added voice chat icons to GUI atlas
 - Improved invalid packet handling on servers
+- Improved resilience against other mods breaking Fabric events